huaping2/requirements.txt
boto3==1.16.21
botocore==1.19.21
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.8
future==0.18.2
idna==2.10
imageio==2.9.0
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.94
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.26.2